pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "tmachine"
version = "0.1.0"
description = "Agnostic 3D Gaussian Splat mutation engine: render → diff → backprop."
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "gsplat>=1.0.0",
    "plyfile>=0.9",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "httpx>=0.27.0",          # TestClient transport
]

# pip install -e ".[ai]"
ai = [
    "diffusers>=0.27.0",
    "transformers>=4.38.0",
    "accelerate>=0.27.0",
    "openai-whisper>=20231117",
    "openai>=1.0.0",
]

# pip install -e ".[api]"
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.9",
    "filelock>=3.13.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
]

# pip install -e ".[full]"
full = ["tmachine[ai,api]"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths   = ["tests"]
addopts     = "-v --tb=short"
markers     = [
    "gpu: marks tests that require CUDA (deselect with -m 'not gpu')",
]
